modification: Finalizado ultimos passos de alguns arquivos que eram responsaveis pela função de deletar o usuário e outros tratamentos e correções.

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Database\QueryException;
+use Illuminate\Support\Facades\DB;
 use App\Models\User;
 
 class UserController extends Controller
@@ -32,7 +33,7 @@ class UserController extends Controller
 
         Auth::login($user);
 
-        return redirect('/dashboard');
+        return redirect('/home');
     }
 
     public function showLoginForm()
@@ -46,7 +47,7 @@ class UserController extends Controller
         
         if (Auth::attempt($credentials)) {
             // Authentication passed...
-            return redirect()->route('home');
+            return redirect()->route('home.welcome');
         }
 
         return back()->withErrors(['email' => 'Email ou senha inválidos.']);
@@ -75,7 +76,7 @@ class UserController extends Controller
     public function edit()
     {
         $user = Auth::user();
-        return view('auth.edit', compact('user'));
+        return view('auth.users.edit', compact('user'));
     }
 
     public function update(Request $request)
@@ -98,29 +99,28 @@ class UserController extends Controller
         try {
             $user->save();
         } catch (\Illuminate\Database\QueryException $ex) {
-            // Tratar exceção
-            dd($ex->getMessage());
+            return redirect()->route('user.edit')->with('error', 'Erro ao atualizar perfil. Por favor, tente novamente.');
         }
-
-        return redirect()->route('user.edit')->with('success', 'Perfil atualizado com sucesso!');
-} 
-    
-    public function destroy($id)
-    {
-        $user = User::find($id);
-
-        if (!$user) {
-            return redirect()->route('users.index')->with('error', 'Usuário não encontrado');
-        }
-
-        $user->delete();
-        return redirect()->route('users.index')->with('success', 'Usuário deletado com sucesso');
+        
+        return redirect()->route('user.edit')->with('success', 'Perfil atualizado com sucesso!');        
     }
-
-
-
-
-
-
-
+    
+    public function destroy()
+    {
+        $user = Auth::user();
+    
+        if (!$user) {
+            return redirect()->route('login')->with('error', 'Usuário não encontrado');
+        }
+    
+        // Restrição para usuário de ID 1
+        if ($user->id === 1) {
+            return redirect()->route('login')->with('error', 'O usuário admin não pode ser deletado.');
+        }
+    
+        DB::table('vaga_candidatos')->where('user_id', $user->id)->delete();
+        $user->delete();
+        Auth::logout();
+        return redirect()->route('login')->with('success', 'Usuário deletado com sucesso');
+    }   
 }
